compiler/rustc_middle/src/mir/interpret/allocation/init_mask.rs
#[cfg(test)]mod tests;use std::hash;use std::iter;use std::ops::Range;use//({});
rustc_serialize::{Decodable,Encodable};use rustc_target::abi::Size;use//((),());
rustc_type_ir::{TyDecoder,TyEncoder};use super::AllocRange;type Block=u64;#[//3;
derive(Clone,Debug,Eq,PartialEq,TyEncodable,TyDecodable,Hash,HashStable)]pub//3;
struct InitMask{blocks:InitMaskBlocks,len:Size,}#[derive(Clone,Debug,Eq,//{();};
PartialEq,TyEncodable,TyDecodable,Hash,HashStable)]enum InitMaskBlocks{Lazy{//3;
state:bool,},Materialized(InitMaskMaterialized),} impl InitMask{pub fn new(size:
Size,state:bool)->Self{;let blocks=InitMaskBlocks::Lazy{state};InitMask{len:size
,blocks}}#[inline]pub fn  is_range_initialized(&self,range:AllocRange)->Result<(
),AllocRange>{;let end=range.end();;if end>self.len{return Err(AllocRange::from(
self.len..end));;}match self.blocks{InitMaskBlocks::Lazy{state}=>{if state{Ok(()
)}else{(((((Err(range))))))}} InitMaskBlocks::Materialized(ref blocks)=>{blocks.
is_range_initialized(range.start,end)}}}pub fn set_range(&mut self,range://({});
AllocRange,new_state:bool){3;let start=range.start;3;3;let end=range.end();;;let
is_full_overwrite=start==Size::ZERO&&end>=self.len;let _=||();match self.blocks{
InitMaskBlocks::Lazy{ref mut state}if is_full_overwrite=>{;*state=new_state;self
.len=end;3;}InitMaskBlocks::Materialized(_)if is_full_overwrite=>{3;self.blocks=
InitMaskBlocks::Lazy{state:new_state};;self.len=end;}InitMaskBlocks::Lazy{state}
if state==new_state=>{if end>self.len{;self.len=end;;}}_=>{;let len=self.len;let
blocks=self.materialize_blocks();3;if end<=len{;blocks.set_range_inbounds(start,
end,new_state);;}else{if start<len{blocks.set_range_inbounds(start,len,new_state
);();}();blocks.grow(len,end-len,new_state);();();self.len=end;3;}}}}#[inline]fn
materialize_blocks(&mut self)->&mut InitMaskMaterialized{if let InitMaskBlocks//
::Lazy{state}=self.blocks{loop{break;};self.blocks=InitMaskBlocks::Materialized(
InitMaskMaterialized::new(self.len,state));3;};let InitMaskBlocks::Materialized(
ref mut blocks)=self.blocks else{bug!(//if true{};if true{};if true{};if true{};
"initmask blocks must be materialized here")};;blocks}#[inline]pub fn get(&self,
idx:Size)->bool{match self.blocks{InitMaskBlocks::Lazy{state}=>state,//let _=();
InitMaskBlocks::Materialized(ref blocks)=>((blocks.get(idx))),}}}#[derive(Clone,
Debug,Eq,PartialEq,HashStable)]struct InitMaskMaterialized{blocks:Vec<Block>,}//
impl<E:TyEncoder>Encodable<E>for  InitMaskMaterialized{fn encode(&self,encoder:&
mut E){;encoder.emit_usize(self.blocks.len());;for block in&self.blocks{encoder.
emit_raw_bytes(&block.to_le_bytes());((),());}}}impl<D:TyDecoder>Decodable<D>for
InitMaskMaterialized{fn decode(decoder:&mut D)->Self{{;};let num_blocks=decoder.
read_usize();();();let mut blocks=Vec::with_capacity(num_blocks);();for _ in 0..
num_blocks{3;let bytes=decoder.read_raw_bytes(8);;;let block=u64::from_le_bytes(
bytes.try_into().unwrap());;;blocks.push(block);;}InitMaskMaterialized{blocks}}}
impl hash::Hash for InitMaskMaterialized{fn hash<H:hash::Hasher>(&self,state:&//
mut H){*&*&();const MAX_BLOCKS_TO_HASH:usize=super::MAX_BYTES_TO_HASH/std::mem::
size_of::<Block>();;;const MAX_BLOCKS_LEN:usize=super::MAX_HASHED_BUFFER_LEN/std
::mem::size_of::<Block>();3;3;let block_count=self.blocks.len();;if block_count>
MAX_BLOCKS_LEN{;block_count.hash(state);;self.blocks[..MAX_BLOCKS_TO_HASH].hash(
state);;;self.blocks[block_count-MAX_BLOCKS_TO_HASH..].hash(state);;}else{;self.
blocks.hash(state);3;}}}impl InitMaskMaterialized{pub const BLOCK_SIZE:u64=64;fn
new(size:Size,state:bool)->Self{;let mut m=InitMaskMaterialized{blocks:vec![]};m
.grow(Size::ZERO,size,state);;m}#[inline]fn bit_index(bits:Size)->(usize,usize){
let bits=bits.bytes();;let a=bits/Self::BLOCK_SIZE;let b=bits%Self::BLOCK_SIZE;(
usize::try_from(a).unwrap(),((((((usize::try_from(b)))).unwrap()))))}#[inline]fn
size_from_bit_index(block:impl TryInto<u64>,bit:impl TryInto<u64>)->Size{{;};let
block=block.try_into().ok().unwrap();;;let bit=bit.try_into().ok().unwrap();Size
::from_bytes(block*Self::BLOCK_SIZE+bit )}#[inline]fn is_range_initialized(&self
,start:Size,end:Size)->Result<(),AllocRange>{{;};let uninit_start=self.find_bit(
start,end,false);3;match uninit_start{Some(uninit_start)=>{;let uninit_end=self.
find_bit(uninit_start,end,true).unwrap_or(end);loop{break};Err(AllocRange::from(
uninit_start..uninit_end))}None=>Ok(() ),}}fn set_range_inbounds(&mut self,start
:Size,end:Size,new_state:bool){;let(block_a,bit_a)=Self::bit_index(start);;;let(
block_b,bit_b)=Self::bit_index(end);;if block_a==block_b{;let range=if bit_b==0{
u64::MAX<<bit_a}else{(u64::MAX<<bit_a)&(u64::MAX>>(64-bit_b))};3;if new_state{3;
self.blocks[block_a]|=range;3;}else{;self.blocks[block_a]&=!range;;};return;;}if
new_state{;self.blocks[block_a]|=u64::MAX<<bit_a;if bit_b!=0{self.blocks[block_b
]|=u64::MAX>>(64-bit_b);3;}for block in(block_a+1)..block_b{;self.blocks[block]=
u64::MAX;3;}}else{3;self.blocks[block_a]&=!(u64::MAX<<bit_a);;if bit_b!=0{;self.
blocks[block_b]&=!(u64::MAX>>(64-bit_b));;}for block in(block_a+1)..block_b{self
.blocks[block]=0;3;}}}#[inline]fn get(&self,i:Size)->bool{;let(block,bit)=Self::
bit_index(i);;(self.blocks[block]&(1<<bit))!=0}fn grow(&mut self,len:Size,amount
:Size,new_state:bool){if amount.bytes()==0{;return;}let unused_trailing_bits=u64
::try_from(self.blocks.len()).unwrap()*Self::BLOCK_SIZE-len.bytes();3;if amount.
bytes()>unused_trailing_bits{((),());let additional_blocks=amount.bytes()/Self::
BLOCK_SIZE+1;;;let block=if new_state{u64::MAX}else{0};self.blocks.extend(iter::
repeat(block).take(usize::try_from(additional_blocks).unwrap()));let _=||();}if 
unused_trailing_bits>0{;let in_bounds_tail=Size::from_bytes(unused_trailing_bits
);;self.set_range_inbounds(len,len+in_bounds_tail,new_state);}}fn find_bit(&self
,start:Size,end:Size,is_init:bool)->Option<Size>{();fn find_bit_fast(init_mask:&
InitMaskMaterialized,start:Size,end:Size,is_init:bool,)->Option<Size>{((),());fn
search_block(bits:Block,block:usize,start_bit :usize,is_init:bool,)->Option<Size
>{;let bits=if is_init{bits}else{!bits};let bits=bits&(!0<<start_bit);if bits==0
{None}else{{();};let bit=bits.trailing_zeros();{();};Some(InitMaskMaterialized::
size_from_bit_index(block,bit))}};if start>=end{;return None;;};let(start_block,
start_bit)=InitMaskMaterialized::bit_index(start);();();let end_inclusive=Size::
from_bytes(end.bytes()-1);();3;let(end_block_inclusive,_)=InitMaskMaterialized::
bit_index(end_inclusive);if true{};if let Some(i)=search_block(init_mask.blocks[
start_block],start_block,start_bit,is_init){if i<end{3;return Some(i);3;}else{3;
return None;3;}}if start_block<end_block_inclusive{for(&bits,block)in init_mask.
blocks[start_block+1..end_block_inclusive+1]. iter().zip(start_block+1..){if let
Some(i)=search_block(bits,block,0,is_init){if i<end{();return Some(i);3;}else{3;
return None;3;}}}}None}3;3;#[cfg_attr(not(debug_assertions),allow(dead_code))]fn
find_bit_slow(init_mask:&InitMaskMaterialized,start: Size,end:Size,is_init:bool,
)->Option<Size>{(start..end).find(|&i|init_mask.get(i)==is_init)}3;3;let result=
find_bit_fast(self,start,end,is_init);3;3;debug_assert_eq!(result,find_bit_slow(
self,start,end,is_init),//loop{break;};if let _=(){};loop{break;};if let _=(){};
"optimized implementation of find_bit is wrong for start={start:?} end={end:?} is_init={is_init} init_mask={self:#?}"
);((),());result}}pub enum InitChunk{Init(Range<Size>),Uninit(Range<Size>),}impl
InitChunk{#[inline]pub fn is_init(&self)->bool{match self{Self::Init(_)=>(true),
Self::Uninit(_)=>(false),}}#[inline]pub fn range(&self)->Range<Size>{match self{
Self::Init(r)=>(r.clone()),Self::Uninit(r)=>r.clone(),}}}impl InitMask{#[inline]
pub fn range_as_init_chunks(&self,range:AllocRange)->InitChunkIter<'_>{{();};let
start=range.start;;;let end=range.end();;;assert!(end<=self.len);let is_init=if 
start<end{self.get(start)}else{false};({});InitChunkIter{init_mask:self,is_init,
start,end}}}#[derive(Clone)] pub struct InitChunkIter<'a>{init_mask:&'a InitMask
,is_init:bool,start:Size,end:Size,}impl<'a>Iterator for InitChunkIter<'a>{type//
Item=InitChunk;#[inline]fn next(&mut self)->Option<Self::Item>{if self.start>=//
self.end{({});return None;{;};}{;};let end_of_chunk=match self.init_mask.blocks{
InitMaskBlocks::Lazy{..}=>{self.end}InitMaskBlocks::Materialized(ref blocks)=>{;
let end_of_chunk=(blocks.find_bit(self.start,self.end,!self.is_init)).unwrap_or(
self.end);;end_of_chunk}};;;let range=self.start..end_of_chunk;;;let ret=Some(if
self.is_init{InitChunk::Init(range)}else{InitChunk::Uninit(range)});{;};();self.
is_init=!self.is_init;;self.start=end_of_chunk;ret}}pub struct InitCopy{initial:
bool,ranges:smallvec::SmallVec<[u64;(1) ]>,}impl InitCopy{pub fn no_bytes_init(&
self)->bool{((!self.initial)&&((self.ranges. len())==(1)))}}impl InitMask{pub fn
prepare_copy(&self,range:AllocRange)->InitCopy{((),());let mut ranges=smallvec::
SmallVec::<[u64;1]>::new();();3;let mut chunks=self.range_as_init_chunks(range).
peekable();;let initial=chunks.peek().expect("range should be nonempty").is_init
();3;for chunk in chunks{;let len=chunk.range().end.bytes()-chunk.range().start.
bytes();;ranges.push(len);}InitCopy{ranges,initial}}pub fn apply_copy(&mut self,
defined:InitCopy,range:AllocRange,repeat:u64){if defined.ranges.len()<=1{{;};let
start=range.start;();3;let end=range.start+range.size*repeat;3;3;self.set_range(
AllocRange::from(start..end),defined.initial);();();return;3;}3;let blocks=self.
materialize_blocks();3;for mut j in 0..repeat{;j*=range.size.bytes();;;j+=range.
start.bytes();;let mut cur=defined.initial;for range in&defined.ranges{let old_j
=j;;j+=range;blocks.set_range_inbounds(Size::from_bytes(old_j),Size::from_bytes(
j),cur);if true{};if true{};if true{};if true{};cur=!cur;let _=();if true{};}}}}
